Fix changes from PR and refactor to remove circular imports and make config consistent
