+++ pkg/promtail/promtail.go
@@ -2,9 +2,9 @@ package promtail
 
 import (
 	"github.com/cortexproject/cortex/pkg/util"
+	"github.com/grafana/loki/pkg/promtail/config"
 	"github.com/weaveworks/common/server"
 
-	"github.com/grafana/loki/pkg/promtail/api"
 	"github.com/grafana/loki/pkg/promtail/client"
 	"github.com/grafana/loki/pkg/promtail/positions"
 	"github.com/grafana/loki/pkg/promtail/targets"
@@ -19,7 +19,7 @@ type Promtail struct {
 }
 
 // New makes a new Promtail.
-func New(cfg api.Config) (*Promtail, error) {
+func New(cfg config.Config) (*Promtail, error) {
 	positions, err := positions.New(util.Logger, cfg.PositionsConfig)
 	if err != nil {
 		return nil, err
