+++ pkg/promtail/targets/filetarget_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/go-kit/kit/log/level"
-	"github.com/grafana/loki/pkg/promtail/api"
 	"github.com/grafana/loki/pkg/promtail/positions"
 
 	"github.com/go-kit/kit/log"
@@ -49,7 +48,7 @@ func TestLongSyncDelayStillSavesCorrectPosition(t *testing.T) {
 		messages: make([]string, 0),
 	}
 
-	target, err := NewFileTarget(logger, client, ps, logFile, nil, &api.TargetConfig{
+	target, err := NewFileTarget(logger, client, ps, logFile, nil, &Config{
 		SyncPeriod: 10 * time.Second,
 	})
 	if err != nil {
@@ -144,7 +143,7 @@ func TestWatchEntireDirectory(t *testing.T) {
 		messages: make([]string, 0),
 	}
 
-	target, err := NewFileTarget(logger, client, ps, logFileDir+"*", nil, &api.TargetConfig{
+	target, err := NewFileTarget(logger, client, ps, logFileDir+"*", nil, &Config{
 		SyncPeriod: 10 * time.Second,
 	})
 	if err != nil {
@@ -234,7 +233,7 @@ func TestFileRolls(t *testing.T) {
 		messages: make([]string, 0),
 	}
 
-	target, err := NewFileTarget(logger, client, positions, dirName+"/*.log", nil, &api.TargetConfig{
+	target, err := NewFileTarget(logger, client, positions, dirName+"/*.log", nil, &Config{
 		SyncPeriod: 10 * time.Second,
 	})
 	if err != nil {
@@ -328,7 +327,7 @@ func TestResumesWhereLeftOff(t *testing.T) {
 		messages: make([]string, 0),
 	}
 
-	target, err := NewFileTarget(logger, client, ps, dirName+"/*.log", nil, &api.TargetConfig{
+	target, err := NewFileTarget(logger, client, ps, dirName+"/*.log", nil, &Config{
 		SyncPeriod: 10 * time.Second,
 	})
 	if err != nil {
@@ -365,7 +364,7 @@ func TestResumesWhereLeftOff(t *testing.T) {
 	}
 
 	// Create a new target, keep the same client so we can track what was sent through the handler.
-	target2, err := NewFileTarget(logger, client, ps2, dirName+"/*.log", nil, &api.TargetConfig{
+	target2, err := NewFileTarget(logger, client, ps2, dirName+"/*.log", nil, &Config{
 		SyncPeriod: 10 * time.Second,
 	})
 	if err != nil {
@@ -433,7 +432,7 @@ func TestGlobWithMultipleFiles(t *testing.T) {
 		messages: make([]string, 0),
 	}
 
-	target, err := NewFileTarget(logger, client, ps, dirName+"/*.log", nil, &api.TargetConfig{
+	target, err := NewFileTarget(logger, client, ps, dirName+"/*.log", nil, &Config{
 		SyncPeriod: 10 * time.Second,
 	})
 	if err != nil {
